[2549] Temporary fix to allow to build helloworld on all stacks
The directories path to `run-redtest` script and logs have changed in
this binding. But it has to be in harmony with the package named `afm-rpm-macros`.

Since it can take some time to have the right version of this package on
every stack, this temporary fix localized in the binding specfile allows
to build it everywhere.

Signed-off-by: Armand Bénéteau <armand.beneteau@iot.bzh>