Fix duplicate of version script
With versions of afb-cmake-modules that doesn't overwrite LINK_FLAGS
the version script was set two times in the link command line,
leading to link errors.

Because the version script is implicitely added when LABEL is set to BINDING,
because using LABEL is the official way when using afb-cmake-modules
and because the version script is optional (not using it has no side effect),
the erronous line is removed.